# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCluster Control Plane API
  version: 0.1.0
  description: HTTP contract for the OpenCluster product API and inbound webhook endpoints.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local self-hosted deployment
...
paths:
...
  /api/v1/integration-types/{type}/connect:
    parameters:
      - $ref: '#/components/parameters/IntegrationType'
    post:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: required-for-unsafe-cookie-request
      x-opencluster-idempotency: not-supported
      parameters:
        - $ref: '#/components/parameters/Organization'
      responses:
        '200':
          $ref: '#/components/responses/ConnectStarted'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: startIntegrationConnect
      summary: Start provider authorization
      tags:
        - Integration Types
  /api/v1/integrations/connect/callback:
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: authenticated
      x-opencluster-organization-selector: forbidden
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      responses:
        '200':
          $ref: '#/components/responses/ConnectLanded'
        '302':
          $ref: '#/components/responses/Redirect'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: completeIntegrationConnect
      summary: Complete provider authorization
      tags:
        - Integrations
      parameters:
        - $ref: '#/components/parameters/OAuthCode'
        - $ref: '#/components/parameters/OAuthState'
  /api/v1/integrations:
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/IntegrationSearch'
        - $ref: '#/components/parameters/IntegrationSort'
        - $ref: '#/components/parameters/IntegrationTypeFilter'
        - $ref: '#/components/parameters/RelayFilter'
        - $ref: '#/components/parameters/DisabledFilter'
      responses:
        '200':
          $ref: '#/components/responses/IntegrationList'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: listIntegrations
      summary: List integrations
      tags:
        - Integrations
    post:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: required-for-unsafe-cookie-request
      x-opencluster-idempotency: not-supported
      parameters:
        - $ref: '#/components/parameters/Organization'
      responses:
        '201':
          $ref: '#/components/responses/IntegrationCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      operationId: createIntegration
      summary: Create integration
      tags:
        - Integrations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegrationRequest'
        required: true
  /api/v1/integrations/{integration}:
    parameters:
      - $ref: '#/components/parameters/IntegrationID'
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      parameters:
        - $ref: '#/components/parameters/Organization'
      responses:
        '200':
          $ref: '#/components/responses/Integration'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: getIntegration
      summary: Get integration
      tags:
        - Integrations
    patch:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: required-for-unsafe-cookie-request
      x-opencluster-idempotency: not-supported
      parameters:
        - $ref: '#/components/parameters/Organization'
      responses:
        '200':
          $ref: '#/components/responses/IntegrationOperation'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: updateIntegration
      summary: Update integration
      tags:
        - Integrations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIntegrationRequest'
        required: true
    delete:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: required-for-unsafe-cookie-request
      x-opencluster-idempotency: not-supported
      parameters:
        - $ref: '#/components/parameters/Organization'
      responses:
        '204':
          description: Integration deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteIntegration
      summary: Delete integration
      tags:
        - Integrations
...
  /api/v1/relays/{registration}/integrations:
    parameters:
      - $ref: '#/components/parameters/RelayRegistrationID'
    get:
      security:
        - SessionCookie: []
      x-opencluster-access: privileged
      x-opencluster-organization-selector: required
      x-opencluster-cookie-csrf: not-required
      x-opencluster-idempotency: not-applicable
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          $ref: '#/components/responses/RelayIntegrationList'
        default:
          $ref: '#/components/responses/ErrorResponse'
      operationId: listRelayIntegrations
      summary: List relay integrations
      tags:
        - Relays
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-oc_session
    BootstrapToken:
      type: http
      scheme: bearer
      description: Deployment bootstrap credential; accepted only until bootstrap succeeds.
    WebhookToken:
      type: apiKey
      in: header
      name: X-OpenCluster-Token
      description: Static sender credential. It authenticates the sender and does not sign the body.
    SlackSignature:
      type: apiKey
      in: header
      name: X-Slack-Signature
      description: Slack request signature verified with the configured signing secret.
  parameters:
    Organization:
      name: X-OpenCluster-Organization
      in: header
      required: true
      description: Selects the Organization whose data and membership are authorized.
      schema:
        $ref: '#/components/schemas/OrganizationID'
      example: 7d722f19-1978-4dc1-a176-e9be92b7e154
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque resume position returned in the preceding page's next field.
      schema:
        type: string
        minLength: 1
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IntegrationSearch:
      name: search
      in: query
      required: false
      description: Case-insensitive containment search over the Integration name.
      schema:
        type: string
        minLength: 1
        maxLength: 256
...
    IntegrationSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum:
          - createdAt
          - -createdAt
...
    IntegrationTypeFilter:
      name: type
      in: query
      required: false
      schema:
        type: string
        enum:
          - alertmanager
          - generic_webhook
          - github
          - kubernetes
          - slack
    RelayFilter:
      name: relay
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/UUID'
    DisabledFilter:
      name: disabled
      in: query
      required: false
      schema:
        type: boolean
...
    OAuthCode:
      name: code
      in: query
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 4096
    OAuthState:
      name: state
      in: query
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 4096
...
    IntegrationID:
      name: integration
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
    IntegrationType:
      name: type
      in: path
      required: true
      schema:
        type: string
        pattern: ^[a-z][a-z0-9_]{0,63}$
    RelayRegistrationID:
      name: registration
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
    ErrorResponse:
      description: The operation did not complete.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ConnectStarted:
      description: Provider-owned installation flow started.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConnectStarted'
    ConnectLanded:
      description: Provider-owned installation flow outcome.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConnectLanded'
    IntegrationList:
      description: Configured Integrations.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IntegrationList'
    IntegrationCreated:
      description: Integration created; webhookSecret is present only in this response when applicable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatedIntegration'
    Integration:
      description: Configured Integration.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Integration'
    IntegrationOperation:
      description: Integration after an operation, including its immediate verification diagnostic when one ran.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IntegrationOperation'
...
    RelayIntegrationList:
      description: Integrations served by a Relay.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RelayIntegrationList'
...
    Redirect:
      description: Browser redirect.
      headers:
        Location:
          required: true
          schema:
            type: string
            format: uri
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or rejected credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Membership, permission, Organization selection, or cookie origin was rejected.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found in the selected Organization.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The requested state conflicts with durable state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    OrganizationID:
      type: string
      format: uuid
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 1024
          example: the selected Organization is not available to this caller
        reason:
          type: string
          enum:
            - credential_rejected
            - session_expired
            - session_revoked
        requires:
          type: string
          minLength: 1
          maxLength: 128
...
    ConnectStarted:
      type: object
      additionalProperties: false
      required:
        - authorizationUrl
        - expiresAt
      properties:
        authorizationUrl:
          type: string
          format: uri
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
    ConnectLanded:
      type: object
      additionalProperties: false
      required:
        - connect
      properties:
        connect:
          type: string
          enum:
            - connected
            - refused
            - unproven
            - unverified
            - installation-taken
        integrationId:
          $ref: '#/components/schemas/UUID'
        note:
          type: string
          maxLength: 1024
    SecretReference:
      type: object
      additionalProperties: false
      required:
        - configured
      properties:
        configured:
          type: boolean
    WebhookReference:
      type: object
      additionalProperties: false
      required:
        - url
        - configured
      properties:
        url:
          type: string
          format: uri-reference
        configured:
          type: boolean
    Integration:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - name
        - status
        - disabled
        - configuration
        - toolAvailability
        - createdAt
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        type:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
        name:
          type: string
          minLength: 1
          maxLength: 128
        status:
          oneOf:
            - type: string
              enum:
                - verified
                - failed
            - type: 'null'
        disabled:
          type: boolean
        configuration:
          type: object
          additionalProperties: true
        relayId:
          $ref: '#/components/schemas/UUID'
        webhook:
          $ref: '#/components/schemas/WebhookReference'
        credential:
          $ref: '#/components/schemas/SecretReference'
        verifiedAt:
          $ref: '#/components/schemas/Timestamp'
        inbound:
          type: object
          additionalProperties: false
          required:
            - available
          properties:
            available:
              type: boolean
            reason:
              type: string
              maxLength: 1024
        toolAvailability:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - tool
              - available
            properties:
              tool:
                type: string
              available:
                type: boolean
              reason:
                type: string
                maxLength: 1024
        createdAt:
          $ref: '#/components/schemas/Timestamp'
...
    IntegrationList:
      type: object
      additionalProperties: false
      required:
        - items
        - next
        - total
      properties:
        next:
          oneOf:
            - type: string
              maxLength: 512
            - type: 'null'
        total:
          oneOf:
            - type: integer
              minimum: 0
            - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
    CreateIntegrationRequest:
      type: object
      additionalProperties: false
      required:
        - type
        - name
        - configuration
      properties:
        type:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
        name:
          type: string
          minLength: 1
          maxLength: 128
        configuration:
          type: object
          additionalProperties: true
        relayId:
          oneOf:
            - $ref: '#/components/schemas/UUID'
            - type: 'null'
    UpdateIntegrationRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        configuration:
          type: object
          additionalProperties: true
    CreatedIntegration:
      type: object
      additionalProperties: false
      required:
        - integration
      properties:
        integration:
          $ref: '#/components/schemas/IntegrationOperation'
        webhookSecret:
          type: string
          minLength: 32
          maxLength: 512
          readOnly: true
          example: ocwh_placeholder_generated_once_1234
...
    RelayIntegrationList:
      type: object
      additionalProperties: false
      required:
        - items
        - next
        - total
      properties:
        next:
          oneOf:
            - type: string
              maxLength: 512
            - type: 'null'
        total:
          oneOf:
            - type: integer
              minimum: 0
            - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/RelayIntegration'
...